Reorganise Project Url feature
