+++ src/App.tsx
@@ -4,9 +4,8 @@ import viteLogo from "/vite.svg"
 import { AddTasks } from "@/addingTasks"
 import { PatAuth } from "@/auth/PatAuth"
 import { PersonalAccessTokenContextProvider } from "@/contexts/PersonalAccessTokenContext"
-import { ProjectUrlContextProvider, useProjectUrl } from "@/contexts/ProjectUrlContext"
+import { ProjectUrlContextProvider, useProjectUrl, SelectProjectUrl } from "@/projectUrl"
 import { IfWorkItemIdIsSet, WorkItemIdContextProvider } from "@/contexts/WorkItemIdContext"
-import { SelectProjectUrl } from "@/projectUrl/SelectProjectUrl"
 import { SelectWorkItem } from "@/selectWorkItem/SelectWorkItem"
 import { WorkItemAndItsChildren } from "@/showWorkItems/WorkItemAndItsChildren"
