+++ src/networking/preconfiguredFetchers.ts
@@ -1,6 +1,6 @@
 import { patAuthMiddleware } from "@/auth/middleware"
 import { usePersonalAccessTokenValue } from "@/contexts/PersonalAccessTokenContext"
-import { useProjectUrl } from "@/contexts/ProjectUrlContext"
+import { useProjectUrl } from "@/projectUrl"
 import { useBasicComposableFetcher } from "@/networking/fetcher"
 import { projectUrlMiddleware, apiVersionMiddleware } from "@/networking/middleware"
