+++ src/projectUrl/types.ts
@@ -1,0 +1,7 @@
+export type UrlType = string
+export type UrlSetterType = (projectUrl: UrlType) => void
+
+export type ContextType = {
+    value: UrlType
+    setter: UrlSetterType
+}
